extract MonitorTracker
