+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/ManagedStrategy.kt
@@ -76,8 +76,8 @@ abstract class ManagedStrategy(
     // Detector of loops or hangs (i.e. active locks).
     internal val loopDetector: LoopDetector = LoopDetector(testCfg.hangingDetectionThreshold)
 
-    // Tracker of acquisitions and releases of monitors.
-    private lateinit var monitorTracker: MonitorTracker
+    // Tracker of the monitors.
+    protected abstract val monitorTracker: MonitorTracker
 
     // InvocationResult that was observed by the strategy during the execution (e.g., a deadlock).
     @Volatile
@@ -207,13 +207,13 @@ abstract class ManagedStrategy(
         finished.fill(false)
         isSuspended.fill(false)
         currentActorId.fill(-1)
-        monitorTracker = MonitorTracker(nThreads)
         traceCollector = if (collectTrace) TraceCollector() else null
         suddenInvocationResult = null
         callStackTrace.forEach { it.clear() }
         suspendedFunctionsStack.forEach { it.clear() }
         randoms.forEachIndexed { i, r -> r.setSeed(i + 239L) }
         localObjectManager = LocalObjectManager()
+        monitorTracker.reset()
     }
 
     /**
@@ -635,9 +635,9 @@ abstract class ManagedStrategy(
         // Therefore, we always release the lock in this case,
         // without tracking the event.
         if (suddenInvocationResult != null) return
-        monitorTracker.releaseMonitor(monitor)
+        val iThread = currentThread
+        monitorTracker.releaseMonitor(iThread, monitor)
         if (collectTrace) {
-            val iThread = currentThread
             val tracePoint = MonitorExitTracePoint(
                 iThread = iThread,
                 actorId = currentActorId[iThread],
@@ -719,13 +719,9 @@ abstract class ManagedStrategy(
     }
 
     override fun notify(monitor: Any, codeLocation: Int, notifyAll: Boolean): Unit = runInIgnoredSection {
-        if (notifyAll) {
-            monitorTracker.notifyAll(monitor)
-        } else {
-            monitorTracker.notify(monitor)
-        }
+        val iThread = currentThread
+        monitorTracker.notify(iThread, monitor, notifyAll = notifyAll)
         if (collectTrace) {
-            val iThread = currentThread
             val tracePoint = NotifyTracePoint(
                 iThread = iThread,
                 actorId = currentActorId[iThread],
@@ -1690,121 +1686,6 @@ internal class ManagedStrategyRunner(
     }
 }
 
-
-/**
- * Tracks synchronization operations with monitors (acquire/release, wait/notify) to maintain a set of active threads.
- */
-private class MonitorTracker(nThreads: Int) {
-    // Maintains a set of acquired monitors with an information on which thread
-    // performed the acquisition and the reentrancy depth.
-    private val acquiredMonitors = IdentityHashMap<Any, MonitorAcquiringInfo>()
-
-    // Maintains a set of monitors on which each thread is waiting.
-    // Note, that a thread can wait on a free monitor if it is waiting for a `notify` call.
-    // Stores `null` if thread is not waiting on any monitor.
-    private val waitingMonitor = Array<MonitorAcquiringInfo?>(nThreads) { null }
-
-    // Stores `true` for the threads which are waiting for a
-    // `notify` call on the monitor stored in `acquiringMonitor`.
-    private val waitForNotify = BooleanArray(nThreads) { false }
-
-    /**
-     * Performs a logical acquisition.
-     */
-    fun acquireMonitor(iThread: Int, monitor: Any): Boolean {
-        // Increment the reentrant depth and store the
-        // acquisition info if needed.
-        val info = acquiredMonitors.computeIfAbsent(monitor) {
-            MonitorAcquiringInfo(monitor, iThread, 0)
-        }
-        if (info.iThread != iThread) {
-            waitingMonitor[iThread] = MonitorAcquiringInfo(monitor, iThread, 0)
-            return false
-        }
-        info.timesAcquired++
-        waitingMonitor[iThread] = null
-        return true
-    }
-
-    /**
-     * Performs a logical release.
-     */
-    fun releaseMonitor(monitor: Any) {
-        // Decrement the reentrancy depth and remove the acquisition info
-        // if the monitor becomes free to acquire by another thread.
-        val info = acquiredMonitors[monitor]!!
-        info.timesAcquired--
-        if (info.timesAcquired == 0)
-            acquiredMonitors.remove(monitor)
-    }
-
-    /**
-     * Returns `true` if the corresponding threads is waiting on some monitor.
-     */
-    fun isWaiting(iThread: Int): Boolean {
-        val monitor = waitingMonitor[iThread]?.monitor ?: return false
-        return waitForNotify[iThread] || !canAcquireMonitor(iThread, monitor)
-    }
-
-    /**
-     * Returns `true` if the monitor is already acquired by
-     * the thread [iThread], or if this monitor is free to acquire.
-     */
-    private fun canAcquireMonitor(iThread: Int, monitor: Any) =
-        acquiredMonitors[monitor]?.iThread?.equals(iThread) ?: true
-
-    /**
-     * Performs a logical wait, [isWaiting] for the specified thread
-     * returns `true` until the corresponding [notify] or [notifyAll] is invoked.
-     */
-    fun waitOnMonitor(iThread: Int, monitor: Any): Boolean {
-        // TODO: we can add spurious wakeups here
-        var info = acquiredMonitors[monitor]
-        if (info != null) {
-            // in case when lock is currently acquired by another thread continue waiting
-            if (info.iThread != iThread)
-                return true
-            // in case when current thread owns the lock we release it
-            // in order to give other thread a chance to acquire it
-            // and put the current thread into waiting state
-            waitForNotify[iThread] = true
-            waitingMonitor[iThread] = info
-            acquiredMonitors.remove(monitor)
-            return true
-        }
-        // otherwise the lock is held by no-one and can be acquired
-        info = waitingMonitor[iThread]
-        check(info != null && info.monitor === monitor && info.iThread == iThread) {
-            "Monitor should have been acquired by this thread"
-        }
-        // if there has been no `notify` yet continue waiting
-        if (waitForNotify[iThread])
-            return true
-        // otherwise acquire monitor restoring its re-entrance depth
-        acquiredMonitors[monitor] = info
-        waitingMonitor[iThread] = null
-        return false
-    }
-
-    /**
-     * Just notify all thread. Odd threads will have a spurious wakeup
-     */
-    fun notify(monitor: Any) = notifyAll(monitor)
-
-    /**
-     * Performs the logical `notifyAll`.
-     */
-    fun notifyAll(monitor: Any): Unit = waitingMonitor.forEachIndexed { iThread, info ->
-        if (monitor === info?.monitor)
-            waitForNotify[iThread] = false
-    }
-
-    /**
-     * Stores the [monitor], id of the thread acquired the monitor [iThread],
-     * and the number of reentrant acquisitions [timesAcquired].
-     */
-    private class MonitorAcquiringInfo(val monitor: Any, val iThread: Int, var timesAcquired: Int)
-}
 
 /**
  * This exception is used to finish the execution correctly for managed strategies.
