+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/modelchecking/ModelCheckingMonitorTracker.kt
@@ -1,0 +1,133 @@
+/*
+ * Lincheck
+ *
+ * Copyright (C) 2019 - 2024 JetBrains s.r.o.
+ *
+ * This Source Code Form is subject to the terms of the
+ * Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
+ * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+package org.jetbrains.kotlinx.lincheck.strategy.managed.modelchecking
+
+import org.jetbrains.kotlinx.lincheck.strategy.managed.*
+import java.util.*
+
+/**
+ * Tracks synchronization operations on the monitors (intrinsic locks)
+ */
+internal class ModelCheckingMonitorTracker(nThreads: Int) : MonitorTracker {
+    // Maintains a set of acquired monitors with an information on which thread
+    // performed the acquisition and the reentrancy depth.
+    private val acquiredMonitors = IdentityHashMap<Any, MonitorAcquiringInfo>()
+
+    // Maintains a set of monitors on which each thread is waiting.
+    // Note, that a thread can wait on a free monitor if it is waiting for a `notify` call.
+    // Stores `null` if thread is not waiting on any monitor.
+    private val waitingMonitor = Array<MonitorAcquiringInfo?>(nThreads) { null }
+
+    // Stores `true` for the threads which are waiting for a
+    // `notify` call on the monitor stored in `acquiringMonitor`.
+    private val waitForNotify = BooleanArray(nThreads) { false }
+
+    /**
+     * Performs a logical acquisition.
+     */
+    override fun acquireMonitor(iThread: Int, monitor: Any): Boolean {
+        // Increment the reentrant depth and store the
+        // acquisition info if needed.
+        val info = acquiredMonitors.computeIfAbsent(monitor) {
+            MonitorAcquiringInfo(monitor, iThread, 0)
+        }
+        if (info.iThread != iThread) {
+            waitingMonitor[iThread] = MonitorAcquiringInfo(monitor, iThread, 0)
+            return false
+        }
+        info.timesAcquired++
+        waitingMonitor[iThread] = null
+        return true
+    }
+
+    /**
+     * Performs a logical release.
+     */
+    override fun releaseMonitor(iThread: Int, monitor: Any) {
+        // Decrement the reentrancy depth and remove the acquisition info
+        // if the monitor becomes free to acquire by another thread.
+        val info = acquiredMonitors[monitor]!!
+        info.timesAcquired--
+        if (info.timesAcquired == 0)
+            acquiredMonitors.remove(monitor)
+    }
+
+    /**
+     * Returns `true` if the corresponding threads is waiting on some monitor.
+     */
+    override fun isWaiting(iThread: Int): Boolean {
+        val monitor = waitingMonitor[iThread]?.monitor ?: return false
+        return waitForNotify[iThread] || !canAcquireMonitor(iThread, monitor)
+    }
+
+    /**
+     * Returns `true` if the monitor is already acquired by
+     * the thread [iThread], or if this monitor is free to acquire.
+     */
+    private fun canAcquireMonitor(iThread: Int, monitor: Any) =
+        acquiredMonitors[monitor]?.iThread?.equals(iThread) ?: true
+
+    /**
+     * Performs a logical wait, [isWaiting] for the specified thread
+     * returns `true` until the corresponding [notify] or [notifyAll] is invoked.
+     */
+    override fun waitOnMonitor(iThread: Int, monitor: Any): Boolean {
+        // TODO: we can add spurious wakeups here
+        var info = acquiredMonitors[monitor]
+        if (info != null) {
+            // in case when lock is currently acquired by another thread continue waiting
+            if (info.iThread != iThread)
+                return true
+            // in case when current thread owns the lock we release it
+            // in order to give other thread a chance to acquire it
+            // and put the current thread into waiting state
+            waitForNotify[iThread] = true
+            waitingMonitor[iThread] = info
+            acquiredMonitors.remove(monitor)
+            return true
+        }
+        // otherwise the lock is held by no-one and can be acquired
+        info = waitingMonitor[iThread]
+        check(info != null && info.monitor === monitor && info.iThread == iThread) {
+            "Monitor should have been acquired by this thread"
+        }
+        // if there has been no `notify` yet continue waiting
+        if (waitForNotify[iThread])
+            return true
+        // otherwise acquire monitor restoring its re-entrance depth
+        acquiredMonitors[monitor] = info
+        waitingMonitor[iThread] = null
+        return false
+    }
+
+    /**
+     * Performs the logical `notify`.
+     * Always notifies all threads, odd threads will simply have a spurious wakeup.
+     */
+    override fun notify(iThread: Int, monitor: Any, notifyAll: Boolean) {
+        waitingMonitor.forEachIndexed { tid, info ->
+            if (monitor === info?.monitor)
+                waitForNotify[tid] = false
+        }
+    }
+
+    override fun reset() {
+        acquiredMonitors.clear()
+        waitingMonitor.fill(null)
+        waitForNotify.fill(false)
+    }
+
+    /**
+     * Stores the [monitor], id of the thread acquired the monitor [iThread],
+     * and the number of reentrant acquisitions [timesAcquired].
+     */
+    private class MonitorAcquiringInfo(val monitor: Any, val iThread: Int, var timesAcquired: Int)
+}
